requirements.txt
torch==2.5.1
torchaudio==2.5.1
streamlit==1.41.1
pillow==11.1.0
torchvision==0.20.1
numpy==2.0.2
retina-face==0.0.17
opencv-python-headless
tf-keras